requirements.txt
pyyaml==6.0.1
aiohttp==3.10.5
aiogram==3.12.0
asynciolimiter==1.0.0
python-dotenv==1.0.0
pillow==10.0.1
cacheout==0.14.1
pydantic==2.4.2
the_retry==0.1.1
webcolors==1.13
dynamic-yaml==2.0.0
motor==3.4.0
numpy==2.0.0
pytz==2024.1
bson==0.5.10
